requirements.txt
pandas
numpy
matplotlib
seaborn
scipy
statsmodels
sklearn
python-pptx
flask
tqdm
pyyaml